renamed paraglide-js plugin into m-function-matcher plugin
