+++ inlang/source-code/marketplace-registry/registry.json
@@ -12,7 +12,6 @@
 		"3i8bor92": "../plugins/i18next/marketplace-manifest.json",
 		"ig84ng0o": "../plugins/json/marketplace-manifest.json",
 		"reootnfj": "../plugins/inlang-message-format/marketplace-manifest.json",
-		"38n9gnir": "../paraglide/paraglide-js-plugin/marketplace-manifest.json",
 		"8y8sxj09": "../versioned-interfaces/language-tag/marketplace-manifest.json",
 		"ezdlll4o": "../versioned-interfaces/translatable/marketplace-manifest.json",
 		"gerre34r": "../paraglide/paraglide-js/marketplace-manifest.json"
