--- inlang/source-code/paraglide/paraglide-js-plugin/CHANGELOG.md
@@ -1,28 +1,0 @@
-# @inlang/plugin-paraglide-js
-
-## 1.3.0
-
-### Minor Changes
-
-- a39638334: add support for new document selector typescriptreact
-- 608ffc963: Refactor message options callback in ideExtensionConfig
-
-## 1.2.0
-
-### Minor Changes
-
-- 6531d0cc3: fix parser to detect correct strings
-
-## 1.1.0
-
-### Minor Changes
-
-- d40e5731a: update paraglide parser
-
-## 1.0.0
-
-### Major Changes
-
-- d0f7578a7: This plugin provides the integration for paraglide-js into the inlang VSCode extension. It allows you to extract messages from your code, view them inline in the text editor and get lints for your messages.
-
-  [README](./README.md)
