--- inlang/source-code/paraglide/paraglide-js-plugin/marketplace-manifest.json
@@ -1,19 +1,0 @@
-{
-	"$schema": "https://inlang.com/schema/marketplace-manifest",
-	"id": "plugin.inlang.paraglide-js",
-	"icon": "https://cdn.jsdelivr.net/gh/inlang/monorepo@main/inlang/source-code/ide-extension/assets/icon.png",
-	"displayName": {
-		"en": "paraglideJS plugin"
-	},
-	"description": {
-		"en": "Enables the inlang VSCode extension to work with paraglide"
-	},
-	"readme": {
-		"en": "https://cdn.jsdelivr.net/gh/inlang/monorepo@main/inlang/source-code/paraglide/paraglide-js-plugin/README.md"
-	},
-	"keywords": ["application", "website", "vscode", "react", "nextjs", "sveltekit", "svelte", "vue"],
-	"publisherName": "inlang",
-	"publisherIcon": "https://inlang.com/favicon/safari-pinned-tab.svg",
-	"license": "Apache-2.0",
-	"module": "https://cdn.jsdelivr.net/npm/@inlang/plugin-paraglide-js@latest/dist/index.js"
-}
