--- inlang/source-code/paraglide/paraglide-js-plugin/src/index.ts
@@ -1,2 +1,0 @@
-import { plugin } from "./plugin.js"
-export default plugin
