--- inlang/source-code/paraglide/paraglide-js-plugin/src/settings.ts
@@ -1,3 +1,0 @@
-export type PluginSettings = object
-
-export const PluginSettings = {}
