--- inlang/source-code/paraglide/paraglide-js-plugin/tsconfig.json
@@ -1,7 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"include": ["src/**/*"],
-	"compilerOptions": {
-		"resolveJsonModule": true
-	}
-}
