+++ inlang/source-code/plugins/m-function-matcher/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @inlang/plugin-m-function-matcher
