+++ inlang/source-code/plugins/m-function-matcher/package.json
@@ -1,0 +1,37 @@
+{
+	"name": "@inlang/plugin-m-function-matcher",
+	"version": "0.1.0",
+	"type": "module",
+	"exports": {
+		".": "./dist/index.js"
+	},
+	"files": [
+		"./dist"
+	],
+	"publishConfig": {
+		"access": "public"
+	},
+	"scripts": {
+		"postinstall": "patch-package --error-on-fail",
+		"dev": "inlang module build --entry ./src/index.ts --outdir ./dist --watch",
+		"build": "inlang module build --entry ./src/index.ts --outdir ./dist",
+		"test": "tsc --noEmit && vitest run --passWithNoTests --coverage",
+		"lint": "eslint ./src --fix",
+		"format": "prettier ./src --write",
+		"clean": "rm -rf ./dist ./.turbo ./node_modules"
+	},
+	"dependencies": {
+		"@inlang/sdk": "*"
+	},
+	"devDependencies": {
+		"@inlang/cli": "*",
+		"@lix-js/fs": "*",
+		"@size-limit/preset-small-lib": "^8.2.4",
+		"@types/parsimmon": "1.10.6",
+		"@vitest/coverage-v8": "^0.33.0",
+		"parsimmon": "^1.18.1",
+		"patch-package": "6.5.1",
+		"typescript": "^5.1.3",
+		"vitest": "0.33.0"
+	}
+}
