+++ inlang/source-code/plugins/m-function-matcher/src/settings.ts
@@ -1,0 +1,3 @@
+export type PluginSettings = object
+
+export const PluginSettings = {}
