+++ inlang/source-code/plugins/t-function-matcher/marketplace-manifest.json
@@ -1,9 +1,9 @@
 {
 	"$schema": "https://inlang.com/schema/marketplace-manifest",
-	"id": "plugin.inlang.t-function-matcher",
+	"id": "plugin.inlang.tFunctionMatcher",
 	"icon": "https://cdn.jsdelivr.net/gh/inlang/monorepo@main/inlang/source-code/ide-extension/assets/icon.png",
 	"displayName": {
-		"en": "t-function matcher"
+		"en": "t-function-matcher Plugin"
 	},
 	"description": {
 		"en": "Enables the inlang VSCode extension to work with t-functions"
